Factor TestStreamSearch out of test/streaming_bisect
